feature
long type converted to bigDecimal

+++ src/main/java/MSGUserService/models/requests/ValidationRequest.java
@@ -5,13 +5,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class ValidationRequest {
 
-    private Long userCode;
+    private BigDecimal userCode;
 
     private String token;
